Refactor: 로그인, 회원가입 완료 버튼을 별도의 컴포넌트로 분리
버튼 안에 이미지가 있어서 복잡해지는 것 같아 별도의 파일로 분리하여 코드 가독성을 높임
