+++ src/pages/Auth/Login/LoginButton.module.css
@@ -9,6 +9,7 @@
 
 .button_text {
     margin: auto 1.5rem;
+    font-weight: var(--font-weight-semibold);
 }
 
 .arrow_img {
