+++ src/pages/Auth/Register/DoneButton.module.css
@@ -5,6 +5,7 @@
 
 .button_text {
     margin: auto 1.5rem;
+    font-weight: var(--font-weight-semibold);
 }
 
 .arrow_img {
