+++ src/pages/Auth/Register/DonePage.tsx
@@ -1,32 +1,11 @@
-import { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
 import AuthLayout from "../../../layouts/AuthLayout/AuthLayout";
 import BaseLayout from "../../../layouts/BaseLayout/BaseLayout";
-import styles from "./DonePage.module.css";
 
 //component
 import AuthBanner from "../AuthBanner";
-import Button from "../../../components/Button/Button";
+import DoneButton from "./DoneButton";
 
 const DonePage = () => {
-  const navigate = useNavigate();
-  const [theme, setTheme] = useState<"light" | "dark">("light");
-
-  // 현재 html에 설정된 theme 속성 읽기
-  useEffect(() => {
-    const currentTheme = document.documentElement.getAttribute("theme");
-    if (currentTheme === "dark") {
-      setTheme("dark");
-    } else {
-      setTheme("light");
-    }
-  }, []);
-
-  // Start 버튼 클릭 시 메인 페이지로 이동
-  const handleNext = () => {
-    navigate("/");
-  };
-
   return (
     <BaseLayout>
       <AuthLayout>
@@ -40,19 +19,7 @@ const DonePage = () => {
         />
 
         {/* 다음 버튼 */}
-        <Button onClick={handleNext}>
-          <div className={styles.inner}>
-            <span className={styles.button_text}>Start</span>
-            <img
-              className={styles.arrow_img}
-              src={
-                theme === "dark"
-                  ? "/arrow_next_white.png"
-                  : "/arrow_next_black.png"
-              }
-            />
-          </div>
-        </Button>
+        <DoneButton />
       </AuthLayout>
     </BaseLayout>
   );
